Added function arguments to scope
