+++ tests/test_ast.py
@@ -1,10 +1,12 @@
 """Xonsh AST tests."""
+import ast as pyast
+
 from xonsh import ast
 from xonsh.ast import Tuple, Name, Store, min_line
 
 import pytest
 
-from tools import  check_parse
+from tools import check_parse, nodes_equal
 
 
 @pytest.fixture(autouse=True)
@@ -31,3 +33,40 @@ def test_multilline_num():
     tree = check_parse(code)
     lsnode = tree.body[1]
     assert 2 == min_line(lsnode)
+
+
+@pytest.mark.parametrize('inp', [
+"""def f():
+    if True:
+        pass
+""",
+"""def f(x):
+    if x:
+        pass
+""",
+"""def f(*args):
+    if not args:
+        pass
+""",
+"""def f(*, y):
+    if y:
+        pass
+""",
+"""def f(**kwargs):
+    if not kwargs:
+        pass
+""",
+"""def f(k=42):
+    if not k:
+        pass
+""",
+"""def f(k=10, *, a, b=1, **kw):
+    if not kw and b:
+        pass
+""",
+])
+def test_args_in_scope(inp):
+    # Context sensitive parsing should not modify AST
+    exp = pyast.parse(inp)
+    obs = check_parse(inp)
+    assert nodes_equal(exp, obs)
