+++ xonsh/ast.py
@@ -431,6 +431,13 @@ class CtxAwareTransformer(NodeTransformer):
         """Handle visiting a function definition."""
         self.ctxadd(node.name)
         self.contexts.append(set())
+        args = node.args
+        argchain = [args.args, args.kwonlyargs]
+        if args.vararg is not None:
+            argchain.append((args.vararg,))
+        if args.kwarg is not None:
+            argchain.append((args.kwarg,))
+        self.ctxupdate(a.arg for a in itertools.chain.from_iterable(argchain))
         self.generic_visit(node)
         self.contexts.pop()
         return node
